Fix for missing toast notifications
